Remove specs interface from Faststream and AsgiFaststream
