+++ faststream/asgi/app.py
@@ -93,18 +93,6 @@ class AsgiFastStream(Application):
         # regular broker args
         logger: Optional["LoggerProto"] = logger,
         lifespan: Optional["Lifespan"] = None,
-        # AsyncAPI args,
-        title: str = "FastStream",
-        version: str = "0.1.0",
-        description: str = "",
-        terms_of_service: Optional["AnyHttpUrl"] = None,
-        license: Optional[Union["License", "LicenseDict", "AnyDict"]] = None,
-        contact: Optional[Union["Contact", "ContactDict", "AnyDict"]] = None,
-        tags: Optional[Sequence[Union["Tag", "TagDict", "AnyDict"]]] = None,
-        external_docs: Optional[
-            Union["ExternalDocs", "ExternalDocsDict", "AnyDict"]
-        ] = None,
-        identifier: Optional[str] = None,
         # hooks
         on_startup: Sequence["AnyCallable"] = (),
         after_startup: Sequence["AnyCallable"] = (),
@@ -115,15 +103,6 @@ class AsgiFastStream(Application):
             broker=broker,
             logger=logger,
             lifespan=lifespan,
-            title=title,
-            version=version,
-            description=description,
-            terms_of_service=terms_of_service,
-            license=license,
-            contact=contact,
-            tags=tags,
-            external_docs=external_docs,
-            identifier=identifier,
             on_startup=on_startup,
             after_startup=after_startup,
             on_shutdown=on_shutdown,
