+++ faststream/specification/asyncapi/__init__.py
@@ -1,9 +1,11 @@
 """AsyncAPI related functions."""
 
-from faststream.specification.asyncapi.generate import get_app_schema
+from .base import AsyncAPIProto
+from .factory import AsyncAPI
 from faststream.specification.asyncapi.site import get_asyncapi_html
 
 __all__ = (
+    "AsyncAPIProto",
+    "AsyncAPI",
     "get_asyncapi_html",
-    "get_app_schema",
 )
