+++ faststream/specification/asyncapi/base/asyncapi.py
@@ -1,0 +1,16 @@
+from abc import abstractmethod
+from typing import Any, Protocol
+
+
+class AsyncAPIProto(Protocol):
+    @abstractmethod
+    def json(self) -> str:
+        ...
+
+    @abstractmethod
+    def jsonable(self) -> Any:
+        ...
+
+    @abstractmethod
+    def yaml(self) -> str:
+        ...
