+++ faststream/specification/asyncapi/v2_6_0/__init__.py
@@ -1,3 +1,7 @@
+from .asyncapi import AsyncAPI2
 from .generate import get_app_schema
 
-__all__ = ("get_app_schema",)
+__all__ = (
+    "AsyncAPI2",
+    "get_app_schema",
+)
