+++ faststream/specification/asyncapi/v3_0_0/__init__.py
@@ -1,7 +1,7 @@
-from . import schema
+from .asyncapi import AsyncAPI3
 from .generate import get_app_schema
 
 __all__ = (
-    "schema",
+    "AsyncAPI3",
     "get_app_schema",
 )
